refactoring: directories structure.
