+++ src/app/initialization/AppInitializer.tsx
@@ -1,6 +1,6 @@
-import useInitializeApp from '../useInitializeApp';
-import AppLoader from './ui/AppLoader';
-import App from './ui/App';
+import AppLoader from '@/app/initialization/ui/AppLoader';
+import App from '@/app/initialization/ui/App';
+import useInitializeApp from './model/useInitializeApp';
 
 const AppInitializer = () => {
   const { isAuth, isInitialized } = useInitializeApp();
