+++ src/app/initialization/model/useInitializeApp.ts
