+++ src/app/layout/AppLayout.tsx
@@ -1,7 +1,7 @@
 import { BottomNavbar, SideNavbar, TopNavbar } from '@/modules/navigation';
 import useMediaQueryType from '@/shared/hooks/helpers/useMediaQueryType';
 import LogoFull from '@assets/logo-full.svg';
-import Layout from './ui/Layout';
+import Layout from '@/app/layout/ui/Layout';
 
 const AppLayout = () => {
   const { isUpMD } = useMediaQueryType();
