+++ src/app/layout/index.ts
@@ -1,1 +1,1 @@
-export { default as AppLayout } from './components/AppLayout';
+export { default as AppLayout } from './AppLayout';
