+++ src/app/providers/Providers.tsx
@@ -1,7 +1,7 @@
 import { ReactElement } from 'react';
 import { Provider } from 'react-redux';
 import { ThemeWrapper } from '@/modules/theme';
-import { store } from '../../../store';
+import { store } from '../../store';
 
 interface ProvidersProps {
   children: ReactElement;
