+++ src/app/providers/index.ts
@@ -1,1 +1,1 @@
-export { default as Providers } from './components/Providers';
+export { default as Providers } from './Providers';
