+++ src/app/routing/Routing.tsx
@@ -4,8 +4,8 @@ import LogIn from '@/pages/LogIn';
 import Home from '@/pages/Home';
 import Components from '@/_old-version/components-presentation/Components';
 import * as ROUTES from '@/shared/constants/routes';
-import ProtectedRoute from './ProtectedRoute';
-import AUTH_PAGES from '../config';
+import ProtectedRoute from '@/app/routing/ui/ProtectedRoute';
+import AUTH_PAGES from './config';
 
 interface RoutingProps {
   isAuth: boolean;
