+++ src/app/routing/config/index.tsx
@@ -1,6 +1,6 @@
 import { lazy, ReactElement } from 'react';
 import * as ROUTES from '@/shared/constants/routes';
-import RouteSuspense from './components/ui/RouteSuspense';
+import RouteSuspense from '@/app/routing/ui/RouteSuspense';
 
 const Profile = lazy(
   () => import('@/_old-version/components/Profile/ProfileContainer')
